+++ .gitignore
@@ -6,6 +6,7 @@ venv/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
+__pycache__
 
 # Logs and databases
 *.log
